backend/requirements.txt
Flask>=2.3
python-dotenv>=1.0
google-genai>=0.7
supabase>=2.6

